refactor(agent): extract agent-registry/node-API into mixin holder
Move profile publishing, agent registration + keystore persistence,
workspace-key rotation/revocation, publishing-conviction NFT operations
and peer messaging/chat/skill/connect helpers (~1.4k lines) into
dkg-agent-registry.ts. Relocate pending chat handler/ACL fields to base.
